requirements.txt
albumentations==1.3.1
hydra-core==1.3.2
numpy==1.24.2
omegaconf==2.3.0
Pillow==9.0.1
pytorch-lightning==2.1.0
torch==2.3.0
mlflow==2.13.0
